Extended version of the Getting Started tutorial (#99)
This PR expands the Getting Started tutorial, analyzing the SmallDSet2 data set from the manuscript.

As a side effect, it also changes things in our CI and dependencies:

We pin Black formatter version. This is probably a permanent change, so I advise to update your local installations (pip  install -r requirements-dev.txt and pre-commit install).
We pin scVI version as it is currently raising an error related to importing Literal. This is hopefully only a temporary issue.